feat:create cacluable module
